+++ src/data/defaults.js
@@ -20,5 +20,5 @@ const defaultSettings = {
 	'show_notice': true,
 	'show_tags': true,
 	'show_words': true,
-	'version': '1.13.0'
+	'version': '1.14.0'
 }; // Initialize default values.